Change /hosting/pages/Bookshelf/index.html import Change /hosting/pages/feed/index.html import Change /hosting/pages/Profile/index.html import Change /hosting/pages/readbook/index.html import upload /hosting/pages/readbook1/index.html upload /hosting/pages/readbook2/index.html upload /hosting/pages/readbook3/index.html upload /hosting/pages/readbook4/index.html Change /hosting/pages/Search/index.html import Change /hosting/pages/Settings/index.html import Change /hosting/pages/store/index.html import Change /hosting/pages/template/index.html import Change /hosting/src/main.ts import upload /hosting/src/scripts/firebase_profile.js upload /hosting/src/scripts/firebase_setting.js Change /hosting/src/style/readbook.css import Delete pages/Bookshelf/index.html Delete pages/feed/index.html Delete pages/Profile/index.html Delete pages/readbook/index.html Delete pages/Search/index.html Delete pages/Settings/index.html Delete pages/store/index.html Delete pages/template/index.html Delete src/counter.ts Delete src/main-save.ts Delete src/main.ts Delete src/style.css Delete src/typescript.svg Delete src/vite env.d.ts Delete src/img/16_9.png Delete src/img/book_FILL0_wght400_GRAD0_opsz48.svg Delete src/img/Book_template.jpg Delete src/img/face_5_FILL0_wght400_GRAD0_opsz48.svg Delete src/img/favicon.ico Delete src/img/feed_FILL0_wght400_GRAD0_opsz48.svgBook_template.jpg Delete src/img/LibraryBarLogo.jpg Delete src/img/search_FILL0_wght400_GRAD0_opsz48.svg Delete src/img/settings_FILL0_wght400_GRAD0_opsz48.svg Delete src/img/store_FILL0_wght400_GRAD0_opsz48.svg Delete src/img/tune_FILL0_wght400_GRAD0_opsz48.svg Delete src/scripts/bookstrap.bundle.min.ts Delete src/scripts/bookstrap.min.ts Delete src/scripts/firebase.js Delete src/scripts/firebaseprofile.js Delete src/scripts/firebasesetting.js Delete src/scripts/jquery-3.6.4.min.js Delete src/scripts/popper.min.js Delete src/scripts/store.js Delete src/scripts/url_add_user_uid.js Delete src/style/book_detail.css Delete src/style/bookstrap.min.css Delete src/style/readbook.css Delete src/style/style.css
