--- pages/Search/index.html
@@ -1,150 +1,0 @@
-<!DOCTYPE html>
-<html lang="zh-Hant-TW">
-
-<head>
-    <meta charset="UTF-8"/>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-    <title>Search</title>
-
-    <link href="../../src/style/bootstrap.min.css" rel="stylesheet">
-    <script type="module" src="../../src/scripts/bootstrap.bundle.min.js"></script>
-    <script type="module" src="../../src/scripts/popper.min.js"></script>
-    <script type="module" src="../../src/scripts/bootstrap.min.js"></script>
-    <script type="module" src="../../src/scripts/jquery-3.6.4.min.js"></script>
-    <script type="module" src="../../src/scripts/store.js"></script>
-    <script type="module" src="../../src/scripts/url_add_user_uid.js"></script>
-
-    <link rel="icon" type="image/x-icon" href="./img/favicon.ico">
-    <link rel="stylesheet" href="https://unicons.iconscout.com/release/v4.0.0/css/line.css"/>
-    <link href="../../src/style/style.css" rel="stylesheet">
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Noto+Sans+TC&display=swap" rel="stylesheet">
-
-    <!-- import lib -->
-    <script src="https://cdn.tailwindcss.com"></script>
-    <link href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200"
-          rel="stylesheet"/>
-</head>
-
-<body>
-    
-    <aside class="sidebar">
-        <nav>
-            <button id="button_Search">
-
-            </button>
-
-            <button id="button_Feed">
-
-            </button>
-
-            <button id="button_Store">
-
-            </button>
-
-            <button id="button_Profile">
-
-            </button>
-
-            <button id="button_Bookshelf">
-                
-            </button>
-
-            <button id="button_Seeting">
-                
-            </button>
-        </nav>
-    </aside>
-
-
-<div class="container-body">
-        <div class="">
-            <font size="6">搜尋 : 鈴芽之旅 結果</font><br>
-            <p>約有 20 筆 相關資料</p>
-        </div>
-
-    <span>書籍</span>
-    <hr>
-
-    <div class="row">
-        <div class="col-sm">
-            <div class="store-ranking-content">
-                <img src="https://cdn.shopify.com/s/files/1/0613/7030/2681/products/212206efffac7d87f8d2a9dc50933b82_473x637.png?v=1672105246">
-                <span class="store-ranking-conten-span">鈴芽之旅</span>
-                <p>新海城</p>
-            </div>
-        </div>
-        <div class="col-sm">
-            <div class="store-ranking-content">
-                <img src="https://cdn.shopify.com/s/files/1/0613/7030/2681/products/212206efffac7d87f8d2a9dc50933b82_473x637.png?v=1672105246">
-                <span class="store-ranking-conten-span">鈴芽之旅</span>
-                <p>新海城</p>
-            </div>
-        </div>
-        <div class="col-sm">
-            <div class="store-ranking-content">
-                <img src="https://cdn.shopify.com/s/files/1/0613/7030/2681/products/212206efffac7d87f8d2a9dc50933b82_473x637.png?v=1672105246">
-                <span class="store-ranking-conten-span">鈴芽之旅</span>
-                <p>新海城</p>
-            </div>
-        </div>
-        <div class="col-sm">
-            <div class="store-ranking-content">
-                <img src="https://cdn.shopify.com/s/files/1/0613/7030/2681/products/212206efffac7d87f8d2a9dc50933b82_473x637.png?v=1672105246">
-                <span class="store-ranking-conten-span">鈴芽之旅</span>
-                <p>新海城</p>
-            </div>
-        </div>
-        <div class="col-sm">
-            <div class="store-ranking-content">
-                <img src="https://cdn.shopify.com/s/files/1/0613/7030/2681/products/212206efffac7d87f8d2a9dc50933b82_473x637.png?v=1672105246">
-                <span class="store-ranking-conten-span">鈴芽之旅</span>
-                <p>新海城</p>
-            </div>
-        </div>
-        <div class="col-sm">
-            <div class="store-ranking-content">
-                <img src="https://cdn.shopify.com/s/files/1/0613/7030/2681/products/212206efffac7d87f8d2a9dc50933b82_473x637.png?v=1672105246">
-                <span class="store-ranking-conten-span">鈴芽之旅</span>
-                <p>新海城</p>
-            </div>
-        </div>
-    </div>
-
-    <span>帳號</span>
-    <hr>
-
-    <div class="row row-cols-6" style="margin-top: 20px;">
-        <div class="col">
-            <img src="../../img/Avatar.png">
-            <span class="store-ranking-conten-span">王曉明</span>
-        </div>
-        <div class="col">
-            <img src="../../img/Avatar.png">
-            <span class="store-ranking-conten-span">王曉明</span>
-        </div>
-        <div class="col">
-            <img src="../../img/Avatar.png">
-            <span class="store-ranking-conten-span">王曉明</span>
-        </div>
-        <div class="col">
-            <img src="../../img/Avatar.png">
-            <span class="store-ranking-conten-span">王曉明</span>
-        </div>
-        <div class="col">
-            <img src="../../img/Avatar.png">
-            <span class="store-ranking-conten-span">王曉明</span>
-        </div>
-        <div class="col">
-            <img src="../../img/Avatar.png">
-            <span class="store-ranking-conten-span">王曉明</span>
-        </div>
-    </div>
-</div>
-
-        
-</body>
-
-</html>
-
