--- pages/feed/index.html
@@ -1,194 +1,0 @@
-<!DOCTYPE html>
-<html lang="zh-Hant-TW">
-
-<head>
-    <!--  Meta data  -->
-    <meta charset="UTF-8"/>
-    <meta http-equiv="X-UA-Compatible" content="IE=edge"/>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-
-    <!--  網頁標題  -->
-    <title>Library Feed</title>
-
-    <!-- 載入 lib -->
-    <script type="module" src="../../src/scripts/bootstrap.bundle.min.js"></script>
-    <script type="module" src="../../src/scripts/popper.min.js"></script>
-    <script type="module" src="../../src/scripts/bootstrap.min.js"></script>
-    <script type="module" src="../../src/scripts/jquery-3.6.4.min.js"></script>
-    <script type="module" src="../../src/scripts/store.js"></script>
-    <script type="module" src="../../src/scripts/url_add_user_uid.js"></script>
-    <script src="https://cdn.tailwindcss.com"></script>
-
-    <link rel="icon" type="image/x-icon" href="./img/favicon.ico">
-    <link rel="stylesheet" href="https://unicons.iconscout.com/release/v4.0.0/css/line.css"/>
-    <link href="../../src/style/bootstrap.min.css" rel="stylesheet">
-    <link href="../../src/style/style.css" rel="stylesheet">
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Noto+Sans+TC&display=swap" rel="stylesheet">
-    <link rel="stylesheet"
-          href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200"/>
-
-</head>
-
-<body>
-
-    
-    <aside class="sidebar">
-        <nav>
-            <button id="button_Search">
-
-            </button>
-
-            <button id="button_Feed">
-
-            </button>
-
-            <button id="button_Store">
-
-            </button>
-
-            <button id="button_Profile">
-
-            </button>
-
-            <button id="button_Bookshelf">
-                
-            </button>
-
-            <button id="button_Seeting">
-                
-            </button>
-        </nav>
-    </aside>
-
-<!-- 網頁右側內容 -->
-<div class="main">
-    <div class="flex flex-col space-y-4">
-        <div class="flex bg-[#D8CCAE] rounded-lg p-4 max-h-96 w-fit h-fit">
-
-            <div class="w-3/4 p-4 grid grid-cols-1 gap-4 content-normal">
-                <div class="text-2xl font-semibold">繼續閱讀：鈴芽之旅</div>
-                <div class="text-lg">150 / 324 頁</div>
-                <div class="bg-gradient-to-b from-gray-500 to-[#D8CCAE] inline-block text-transparent bg-clip-text text-base">
-                    十七歲的少女鈴芽和阿姨住在九州靜謐的海港城鎮。某天上學途中，鈴芽與一名俊美的青年擦身而過。想要前往「廢墟」的青年，說著他在「尋找門」。鈴芽追著他，來到了山中的廢墟，在那裡的只有一扇陳舊的白門，
-                </div>
-
-                <div onclick="location.href='../readbook/'"
-                     class="bg-white hover:bg-slate-200 w-48 h-12 py-2 px-4 rounded-full text-center text-lg flex items-center justify-center cursor-pointer select-none">
-                    閱讀
-                </div>
-            </div>
-
-            <div class="w-1/4 content-center flex items-center justify-center my-2">
-                <img class="object-contain rounded-md w-full md:w-4/5"
-                     src="https://cdn.shopify.com/s/files/1/0613/7030/2681/products/212206efffac7d87f8d2a9dc50933b82_473x637.png?v=1672105246"
-                     alt="Book"/>
-            </div>
-
-        </div>
-
-        <!-- Feed Content -->
-        <div class="flex flex-col bg-white rounded-3xl p-12 gap-4">
-
-            <div class="flex flex-row gap-4">
-                <div>
-                    <img
-                      src="https://storage.googleapis.com/pxcode-pro/644d82b70649a2000afe5491/assets/8b8a920c3d37a86d2cebb867d3453fcc.png"
-                      alt="alt text"
-                    />
-                </div>
-                <div class="flex-1 w-64 ">
-                    <div class="grid grid-cols-1 col-start-2 place-items-start content-end items-center m-2">
-                        <div class="text-lg">
-                            王曉明
-                        </div>
-                        <div class="text-sm">
-                            2021/10/10
-                        </div>
-                    </div>
-                </div>
-
-            </div>
-            <div class="flex flex-row gap-4">
-                <div class="ml-20 flex-1 w-64 text-base col-span-2">
-                    王曉明看了 鈴芽之旅 已觀看 50 %
-                </div>
-            </div>
-            <div class="bg-[#7EB3BF] rounded-lg col-span-2 grid grid-cols-3 gap-4 px-6 py-4 ml-20">
-                <img class="object-contain rounded-lg m-4"
-                     src="https://cdn.shopify.com/s/files/1/0613/7030/2681/products/212206efffac7d87f8d2a9dc50933b82_473x637.png?v=1672105246"
-                     alt="Book"/>
-                <div class="content-normal">
-                    <p1 class="text-[rgb(242,242,242)]">鈴芽之旅</p1>
-                    <br><br>
-                    <u class="text-[rgb(242,242,242)] text-sm">新海 誠</u>
-                    <br><br>
-                    <div class="h-40 bg-gradient-to-b from-gray-500 inline-block text-transparent bg-clip-text text-justify">
-                        <p>十七歲的少女鈴芽和阿姨住在九州靜謐的海港城鎮。某天上學途中，鈴芽與一名俊美的青年擦身而過。想要前往「廢墟」的青年，說著他在「尋找門」。鈴芽追著他，來到了山中的廢墟，在那裡的只有一扇陳舊的白門，</p>
-                    </div>
-                    <br>
-                    <div class="grid place-items-end">
-                        <div onclick="location.href='../readbook/'" class="bg-white hover:bg-slate-200 w-24 h-12 py-2 px-4 rounded-full text-black text-center text-base flex items-center justify-center cursor-pointer select-none">
-                            閱讀
-                        </div>
-                    </div>
-                </div>
-            </div>
-
-            <div class="flex flex-row gap-4 ml-20">
-                <div class="flex flex-row gap-4">
-                    <span class="material-symbols-outlined">
-                    favorite
-                    </span>
-                    <div class="text-base">
-                        10
-                    </div>
-                </div>
-                <div class="flex flex-row gap-4">
-                    <span class="material-symbols-outlined">
-                    chat
-                    </span>
-                    <div class="text-base">
-                        10
-                    </div>
-                </div>
-                <div class="flex flex-row gap-4">
-                    <span class="material-symbols-outlined">
-                    timelapse
-                    </span>
-                    <div class="text-base">
-                        20 : 30
-                        <span class="text-xs">秒</span>
-                    </div>
-                </div>
-                <div class="flex flex-row gap-4 ml-4">
-                    <div>
-                        <img
-                          src="https://storage.googleapis.com/pxcode-pro/644d82b70649a2000afe5491/assets/8b8a920c3d37a86d2cebb867d3453fcc.png"
-                          alt="alt text"
-                          class="w-6 h-6"
-                        />
-                    </div>
-                    <div class="text-sm">
-                        李大華 : 哇~ 太酷了吧！
-                    </div>
-                    <button onclick=" window.open('https://www.instagram.com/','_blank')" class="material-symbols-outlined h-3 w-3">
-                        reply
-                    </button>
-                </div>
-
-            </div>
-
-
-        </div>
-
-
-    </div>
-
-</div>
-
-
-</body>
-
-</html>
