--- pages/store/index.html
@@ -1,109 +1,0 @@
-<!DOCTYPE html>
-<html lang="zh-Hant-TW">
-
-<head>
-    <meta charset="UTF-8"/>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-    <title>Store</title>
-
-    <link href="../../src/style/bootstrap.min.css" rel="stylesheet">
-    <script type="module" src="../../src/scripts/bootstrap.bundle.min.js"></script>
-    <script type="module" src="../../src/scripts/popper.min.js"></script>
-    <script type="module" src="../../src/scripts/bootstrap.min.js"></script>
-    <script type="module" src="../../src/scripts/jquery-3.6.4.min.js"></script>
-    <script type="module" src="../../src/scripts/store.js"></script>
-    <script type="module" src="../../src/scripts/url_add_user_uid.js"></script>
-
-    <link rel="icon" type="image/x-icon" href="./img/favicon.ico">
-    <link rel="stylesheet" href="https://unicons.iconscout.com/release/v4.0.0/css/line.css"/>
-    <link href="../../src/style/style.css" rel="stylesheet">
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Noto+Sans+TC&display=swap" rel="stylesheet">
-
-    <!-- import lib -->
-    <script src="https://cdn.tailwindcss.com"></script>
-    <link href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200"
-          rel="stylesheet"/>
-</head>
-
-<body>
-    <script src="https://www.gstatic.com/firebasejs/8.3.0/firebase-app.js"></script>
-    <script src="https://www.gstatic.com/firebasejs/8.3.0/firebase-firestore.js"></script>
-    <script src="../../src/scripts/firebase.js"></script>
-    <aside class="sidebar">
-        <nav>
-            <button id="button_Search">
-
-            </button>
-
-            <button id="button_Feed">
-
-            </button>
-
-            <button id="button_Store">
-
-            </button>
-
-            <button id="button_Profile">
-
-            </button>
-
-            <button id="button_Bookshelf">
-                
-            </button>
-
-            <button id="button_Seeting">
-                
-            </button>
-        </nav>
-    </aside>
-
-
-<div class="container-body">
-    <div class="recommend-carousel">
-        <div class="recommend-carousel-inner" id="recommend_book">
-            <!--自動加載-->
-        </div>
-    </div>
-
-    <span>綜合排名</span>
-    <hr>
-
-    <div class="row row-cols-sm-1 row-cols-md-2 row-cols-lg-3 row-cols-xl-6 " id="store_ranking">
-        <!--自動加載-->
-    </div>
-
-    <span>分類</span>
-    <hr>
-
-    <div class="row row-cols-md-2 row-cols-lg-4 row-cols-xl-6 " id="store_class">
-        <!--自動加載-->
-    </div>
-</div>
-
-<!-- book_detail -->
-<div id="page_book_detail0" class="page_book_detail">
-    <!--自動加載-->
-</div>
-<div id="page_book_detail1" class="page_book_detail">
-    <!--自動加載-->
-</div>
-<div id="page_book_detail2" class="page_book_detail">
-    <!--自動加載-->
-</div>
-<div id="page_book_detail3" class="page_book_detail">
-    <!--自動加載-->
-</div>
-<div id="page_book_detail4" class="page_book_detail">
-    <!--自動加載-->
-</div>
-<div id="page_book_detail5" class="page_book_detail">
-    <!--自動加載-->
-</div>
-
-</body>
-
-</div>
-
-</html>
