--- pages/template/index.html
@@ -1,70 +1,0 @@
-<!DOCTYPE html>
-<html lang="zh-Hant-TW">
-
-<head>
-    <!--  Meta data  -->
-    <meta charset="UTF-8"/>
-    <meta http-equiv="X-UA-Compatible" content="IE=edge"/>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-
-    <!--  網頁標題  -->
-    <title>Library Bar</title>
-
-    <!-- 載入 lib -->
-    <script type="module" src="../../src/scripts/bootstrap.bundle.min.js"></script>
-    <script type="module" src="../../src/scripts/popper.min.js"></script>
-    <script type="module" src="../../src/scripts/bootstrap.min.js"></script>
-    <script type="module" src="../../src/scripts/jquery-3.6.4.min.js"></script>
-    <script type="module" src="../../src/scripts/store.js"></script>
-    <script src="https://cdn.tailwindcss.com"></script>
-    <script type="module" src="../../src/scripts/url_add_user_uid.js"></script>
-
-    <link rel="icon" type="image/x-icon" href="./img/favicon.ico">
-    <link rel="stylesheet" href="https://unicons.iconscout.com/release/v4.0.0/css/line.css"/>
-    <link href="../../src/style/bootstrap.min.css" rel="stylesheet">
-    <link href="../../src/style/style.css" rel="stylesheet">
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Noto+Sans+TC&display=swap" rel="stylesheet">
-    <link rel="stylesheet"
-          href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200"/>
-
-</head>
-
-<body>
-  
-    
-    <aside class="sidebar">
-        <nav>
-            <button id="button_Search">
-
-            </button>
-
-            <button id="button_Feed">
-
-            </button>
-
-            <button id="button_Store">
-
-            </button>
-
-            <button id="button_Profile">
-
-            </button>
-
-            <button id="button_Bookshelf">
-                
-            </button>
-
-            <button id="button_Seeting">
-                
-            </button>
-        </nav>
-    </aside>
-
-<!-- 網頁右側內容 -->
-<div class="main">
-</div>
-
-</body>
-</html>
